BusBooking/requirements.txt
Django
django-import-export
pip install django-allauth
